% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opls-methods.R
\name{print,opls-method}
\alias{print,opls-method}
\alias{print.opls}
\title{Print method for 'opls' objects}
\usage{
\S4method{print}{opls}(x)
}
\arguments{
\item{x}{An S4 object of class \code{opls}, created by the \code{opls}
function.}
}
\value{
Invisible.
}
\description{
Displays information about the dataset and the model.
}
\examples{

data(sacurine)
attach(sacurine)
sacurine.plsda <- opls(dataMatrix, sampleMetadata[, "gender"])

print(sacurine.plsda)

detach(sacurine)

}
