% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roplsMini-package.R
\docType{package}
\name{roplsMini-package}
\alias{roplsMini-package}
\title{A lightweight package ropls wrapper}
\description{
Removed ExpressionSet, MultiAssayExperiment, MultiDataSet and SummarizedExperiment objects,
only retained PCA, PLS(-DA) and OPLS(-DA) methods for data.frame and matirx in ropls
(https://bioconductor.org/packages/release/bioc/html/ropls.html).
}
\author{
Zimlik
}
\keyword{package}
